refactor: change folder structure (#61)
