+++ src/features/qna-detail/index.ts
@@ -1,0 +1,3 @@
+export { default as QnaDetailHeader } from './components/QnaDetailHeader'
+export { default as QnaDetailAnswer } from './components/QnaDetailAnswer'
+export { default as AnswerCard } from './components/AnswerCard'
