+++ src/features/qna-detail/types.ts
