+++ src/pages/qna-detail/QnaDetailPage.tsx
@@ -1,12 +1,13 @@
 // import { useParams } from 'react-router-dom'
 import { EmptyState } from '@/components'
-import QnaDetailHeader from './components/QnaDetailHeader'
+import { QnaDetailHeader, QnaDetailAnswer } from '@/features/qna-detail'
 import { mockQuestionDetail } from './mock'
-import AnswerCard from './components/QnaDetailAnswer'
 
 export default function QnaDetailPage() {
   // 임시 데이터용 questionId, 실제로는 API에서 받아올 예정
   //   const { questionId } = useParams()
+
+  // TODO: 회원일 경우 질문자와 답변자에 따라 다른 UI 보여주기
 
   // 지금은 mock 데이터 사용
   const question = mockQuestionDetail
@@ -24,10 +25,11 @@ export default function QnaDetailPage() {
       <QnaDetailHeader question={question} onShare={handleShare} />
 
       {/* 답변 카드 */}
-      <AnswerCard answers={question.answers} />
-
-      {/* 답변이 없을 때 */}
-      {question.answers.length === 0 && <EmptyState type="emptyState" />}
+      {question.answers.length > 0 ? (
+        <QnaDetailAnswer answers={question.answers} />
+      ) : (
+        <EmptyState type="emptyState" />
+      )}
     </div>
   )
 }
